start work with code
