+++ src/index.ts
@@ -1,15 +1,18 @@
 import { Todo } from "./todos"
 import { Project} from "./projects"
 import './sass/style.scss'
-import { hookToElements } from "./hook_assets"
-import { readTask } from "./domManager"
+import { DOMManager } from "./domManager"
+
+
 console.log("Hello World")
 
 let todo = new Todo(
     "Write proejct"
 )
-hookToElements()
-
+let dom = new DOMManager();
+dom.hookSVGToElements()
+let result = dom.readTask()
+console.log("Result is: ", result)
 
 
 console.log(todo.info())
