+++ src/templates/index.html
@@ -71,7 +71,7 @@
         <!--The sticky footer -->
         <div class="footer">
             <div id="add-task-container">
-                <img id="add-task"><input type="text" name="task-add" id="task-add" placeholder="Add Task">
+                <img id="add-task"><input type="text" name="task-add" id="task-add" placeholder="Add Task" onkeydown="readTask()">
             </div>
         </div>
     </section>
